Removing unused variables
When running with -Wall options to the compiler, it will complain about
unused variables. Similarly, -Werror will then make these builds fail.
This patch removes any unused variable; most of which seem to be
status codes which are stored but never used, or "old" state variables
that are not static nor stored outside of the function stack and as such
are lost at the end of every function execution.